Factor out "raw" Arrhenius parts
This is so we can reuse them for the vector version of the test.
